integrated sensors with GUI
graph plots light and accel data
can change estop for accel
light comes on at night time 5lux
sensor page shows current reading of sensors
top bar changes image with sensor data

+++ EGH456MotorProject/sensors/opt3001.c
@@ -92,7 +92,7 @@ bool opt3001_read(uint16_t *rawData)
  *
  * @return      none
  **************************************************************************************************/
-void opt3001_convert(uint16_t rawData, float *convertedLux)
+void opt3001_convert(uint16_t rawData, double *convertedLux)
 {
     uint16_t e, m;
